Remove unused "buttons"
Given that the controller is polled twice, once with the SELECT
line set low, and once with the SELECT line set high,
theoretically 12 buttons could be registered, but with the
controller I have, UP and DOWN are wired directly to to the cable
(hence reading the same regardless of the SELECT line), and the
alternate pins for LEFT and RIGHT are wired to ground, so they will
register inactive all the time. Thus, these can be removed from
both the polling and the display.